overlord/exportstate: move handlers to handlers.go
